refactor to DbTestFixture
